:zap: rewrite; much faster and accepts more arguments
The bottleneck is finding all the `node_modules`. A typical `node_modules` folder can easily include several hundred child `node_modules`.

Before, this would continue to search in all child directories, but would not create a file if it was a child `node_modules`. Now, the script no longer searches within the `node_modules`. This makes the script significantly faster to keep the same behavior.

Also, arguments are parsed differently. A new `--version` flag has been added. `--help` no longer has to be the first argument. Unknown flags cause the script to error before running further.

The script should be generally easier to read. Before, the main section used a while loop and process substitution. This made it difficult to read mainly because the input was specified at the end of the loop. Now, the script uses a for-in loop which is much easier to read.